Move component tree
